feat(assistant): 图片点击放大 Lightbox 及 OpenSpec 归档
提取 ImageLightbox 为独立组件供聊天图片和附件缩略图复用，
点击对话中的图片或待发送附件缩略图可全屏预览，支持 Esc 关闭。
同步 delta specs 到主目录并归档 chat-image-input change。
