+++ openspec/changes/archive/2026-03-13-chat-image-input/.openspec.yaml
