+++ openspec/changes/archive/2026-03-13-chat-image-input/specs/chat-image-attachment/spec.md
@@ -1,0 +1,58 @@
+## ADDED Requirements
+
+### Requirement: 图片附件输入
+系统 SHALL 允许用户在对话框输入区通过粘贴、点击上传、拖拽三种方式附加图片。
+
+#### Scenario: 粘贴图片
+- **WHEN** 用户在对话框输入区按 Ctrl+V（或 Cmd+V），剪贴板中包含图片
+- **THEN** 系统将图片添加到附件列表，并在输入框上方显示缩略图
+
+#### Scenario: 点击上传
+- **WHEN** 用户点击输入框旁的附件按钮，从文件选择器选取一张或多张图片
+- **THEN** 系统将所选图片添加到附件列表，并显示缩略图
+
+#### Scenario: 拖拽图片
+- **WHEN** 用户将图片文件拖拽至对话框输入区并释放
+- **THEN** 系统将图片添加到附件列表，并显示缩略图；拖入过程中输入区 SHALL 高亮显示放置反馈
+
+#### Scenario: 超出数量上限
+- **WHEN** 当前附件数量已达 5 张，用户尝试再次添加图片
+- **THEN** 系统忽略新图片，附件按钮 SHALL 变为禁用状态
+
+#### Scenario: 超出文件大小上限
+- **WHEN** 用户添加单张大于 5MB 的图片
+- **THEN** 系统拒绝添加并向用户显示错误提示
+
+#### Scenario: 移除附件
+- **WHEN** 用户点击某张缩略图右上角的删除按钮
+- **THEN** 系统从附件列表移除该图片，缩略图消失
+
+### Requirement: 携带图片发送消息
+系统 SHALL 在用户发送消息时将附件图片与文字内容一并提交给 Agent。
+
+#### Scenario: 发送含图片的消息
+- **WHEN** 用户在附件列表非空时点击发送（或按 Enter）
+- **THEN** 系统将文字与图片 base64 数据组合成 multimodal 消息发送，发送后附件列表 SHALL 清空
+
+#### Scenario: 仅发送文字
+- **WHEN** 用户在附件列表为空时发送消息
+- **THEN** 系统行为与原有纯文本发送保持一致
+
+### Requirement: 图片消息渲染
+系统 SHALL 在对话历史中正确渲染用户发送的图片，并支持点击放大查看。
+
+#### Scenario: 发送后即时显示
+- **WHEN** 含图片的消息发送成功
+- **THEN** 用户气泡中 SHALL 显示图片缩略图（最大高度 256px），文字内容显示在图片下方
+
+#### Scenario: 历史回放显示
+- **WHEN** 用户重新加载已有会话
+- **THEN** 含图片的历史消息 SHALL 正确渲染图片内容
+
+#### Scenario: 点击缩略图放大
+- **WHEN** 用户点击对话中的图片缩略图
+- **THEN** 系统以 lightbox 形式全屏显示原图，点击遮罩或按 Esc 关闭
+
+#### Scenario: 点击待发送图片放大
+- **WHEN** 用户点击输入区附件缩略图
+- **THEN** 系统以 lightbox 形式全屏显示该图片，点击遮罩或按 Esc 关闭
