+++ openspec/specs/chat-image-attachment/spec.md
@@ -39,7 +39,7 @@
 - **THEN** 系统行为与原有纯文本发送保持一致
 
 ### Requirement: 图片消息渲染
-系统 SHALL 在对话历史中正确渲染用户发送的图片。
+系统 SHALL 在对话历史中正确渲染用户发送的图片，并支持点击放大查看。
 
 #### Scenario: 发送后即时显示
 - **WHEN** 含图片的消息发送成功
@@ -48,3 +48,11 @@
 #### Scenario: 历史回放显示
 - **WHEN** 用户重新加载已有会话
 - **THEN** 含图片的历史消息 SHALL 正确渲染图片内容
+
+#### Scenario: 点击缩略图放大
+- **WHEN** 用户点击对话中的图片缩略图
+- **THEN** 系统以 lightbox 形式全屏显示原图，点击遮罩或按 Esc 关闭
+
+#### Scenario: 点击待发送图片放大
+- **WHEN** 用户点击输入区附件缩略图
+- **THEN** 系统以 lightbox 形式全屏显示该图片，点击遮罩或按 Esc 关闭
